# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Anby Wiki API
  version: "0.1.0"
  description: |
    Anby Wiki HTTP API 权威契约（OpenAPI 3.1）。
    本文件是所有接口行为的唯一权威定义：错误模型、分页、幂等头与请求 ID 在此统一定义。
    前端客户端由 OpenAPI Generator 从本文件生成，禁止手写 URL/DTO。
    Go 侧 handler 手写实现，由 Contract Test 保证与本文件一致。
  contact:
    name: Anby Wiki
...
servers:
  - url: http://localhost:3000
    description: 本地 Web 入口（Next.js 将 /api/* 转发到 Go API）
  - url: http://localhost:8080
    description: 本地直连 Go API
...
paths:
...
  /api/v1/pages/{id}:
    get:
      operationId: getPageByID
      summary: 按 ID 读取页面当前版本
      description: |
        阅读端点（匿名可读，无需登录）。按页面 ID 读取当前 Revision 与渲染 HTML，
        响应结构与行为同 getPageByTitle（重定向同样跟随；软删除页返回 410 gone；
        未发布过 content 为 null）。
      tags: [reading]
      parameters:
        - $ref: "#/components/parameters/PageId"
        - $ref: "#/components/parameters/ContentMode"
      responses:
        "200":
          description: 页面元信息 + 当前 Revision 元信息 + AST + 渲染 HTML
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageWithContent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "410":
          $ref: "#/components/responses/Gone"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/Internal"
  /api/v1/pages:
    post:
      operationId: createPage
      summary: 创建页面
      description: |
        在默认站点（wiki 固定为种子里 site_key='default' 的站点）创建页面。
        标题经服务端规范化（NFC、大小写折叠、空白折叠），与同 wiki+namespace 的
        活页面或别名冲突时返回 409。
        Actor 身份由服务端 session cookie 解析；客户端不能声明 Actor。
      tags: [pages]
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePageRequest"
      responses:
        "201":
          description: 页面已创建
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/Internal"
  /api/v1/pages/{id}/rename:
    post:
      operationId: renamePage
      summary: 页面改名
      description: |
        更新页面标题，旧标题写入 page_alias（Page ID 不变）。
        新标题被同 wiki+namespace 的活页面或其他页面的别名占用时返回 409。
      tags: [pages]
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/PageId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenamePageRequest"
      responses:
        "200":
          description: 改名后的页面
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/Internal"
...
  /api/v1/pages/{id}/block-redirects:
    get:
      operationId: listBlockRedirects
      summary: 列出页面的稳定章节迁移映射
      tags: [pages]
      parameters:
        - $ref: "#/components/parameters/PageId"
      responses:
        "200":
          description: BlockRedirect 列表
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BlockRedirectList" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/pages/{id}/block-redirects/{block_id}:
    put:
      operationId: upsertBlockRedirect
      summary: 创建或更新稳定章节迁移映射
      description: 目标会解析并折叠到当前有效 Heading Block；服务端拒绝环和悬空目标。
      tags: [pages]
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/PageId"
        - name: block_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpsertBlockRedirectRequest" }
      responses:
        "200":
          description: 当前 BlockRedirect
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BlockRedirect" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
    delete:
      operationId: deleteBlockRedirect
      summary: 删除错误的章节迁移映射（审计历史保留）
      tags: [pages]
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/PageId"
        - name: block_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: 已删除 }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/pages/{id}/revisions:
    get:
      operationId: listRevisions
      summary: Revision 历史列表
      description: |
        阅读端点（匿名可读，无需登录）。按 (created_at DESC, id DESC) 游标分页
        列出页面 Revision 元信息（含冗余自快照的 content_hash/schema_version，不含 AST）。
        首页不传 cursor；后续页传上一页响应的 next_cursor；next_cursor 为 null 表示没有更多。
        游标无法解析返回 400 validation_failed；页面不存在返回 404。
      tags: [history]
      parameters:
        - $ref: "#/components/parameters/PageId"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: 一页 Revision 历史
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevisionListPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Internal"
    post:
      operationId: publishRevision
      summary: 发布 Revision
      description: |
        原子发布一个 Revision：校验 AST（Typed Block AST v1，schema_version=1）后，
        在单事务内写入 ContentSnapshot / Revision / 页面当前指针 / AuditEvent / OutboxEvent。
        乐观锁：expected_revision_id 必须等于页面当前 Revision（首发布不传，
        要求页面尚无 Revision）；不一致返回 409 stale_revision。
        content_hash 与 size_bytes 由服务端对 canonical AST 计算，客户端提供的值不被信任。
      tags: [pages]
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/PageId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishRevisionRequest"
      responses:
        "201":
          description: 发布成功的 Revision
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Revision"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: 并发冲突（陈旧基线 stale_revision）
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: stale_revision
                message: "page: 基线 Revision 已过期: page=… expected=… current=…"
                request_id: 9f1c2ab3d4e5f607
        "500":
          $ref: "#/components/responses/Internal"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: anby_session
      description: HttpOnly 同源服务端 session cookie；名称可由部署配置覆盖
  headers:
    X-Request-ID:
      description: 请求 ID；客户端可携带同名请求头，服务端透传或重新生成
      schema:
        type: string
  parameters:
    PageId:
      name: id
      in: path
      required: true
      description: 页面 ID（UUIDv7）
      schema:
        type: string
        format: uuid
...
    ContentMode:
      name: content_mode
      in: query
      required: false
      description: >-
        auto 对大于 64 KiB 且分片投影就绪的页面返回 sections；full 强制全文；
        sections 在投影就绪时强制分片，否则安全降级为 full。
      schema:
        type: string
        enum: [auto, full, sections]
        default: auto
...
    Cursor:
      name: cursor
      in: query
      required: false
      description: 上一页响应返回的 next_cursor；首页不传
      schema:
        type: string
    PageSize:
      name: page_size
      in: query
      required: false
      description: 每页条数，默认 20，最大 100
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    Page:
      type: object
      required:
        - id
        - wiki_id
        - namespace_id
        - normalized_title
        - display_title
        - language
        - content_model
        - status
        - current_revision_id
        - created_by
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        wiki_id:
          type: string
          format: uuid
        namespace_id:
          type: string
          format: uuid
        normalized_title:
          type: string
          description: 规范化标题（路由唯一性键之一）
          example: anby demara
        display_title:
          type: string
          description: 展示标题
          example: Anby Demara
        language:
          type: string
          example: zh-Hans
        content_model:
          type: string
          example: block-v1
        status:
          type: string
          example: active
        current_revision_id:
          type: string
          format: uuid
          nullable: true
          description: 当前 Revision；未发布过为 null
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    CreatePageRequest:
      type: object
      required: [namespace, title]
      properties:
        namespace:
          type: string
          description: 命名空间 key（如 main / talk / user / project / component / collection / file）
          example: main
        title:
          type: string
          minLength: 1
          maxLength: 255
          example: Anby Demara
        language:
          type: string
          description: 页面语言，缺省用站点默认语言
          example: zh-Hans
        content_model:
          type: string
          description: 内容模型，缺省 block-v1
          example: block-v1
    RenamePageRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          example: Anby Demara (New)
...
    UpsertBlockRedirectRequest:
      type: object
      additionalProperties: false
      required: [target_page_id, target_block_id]
      properties:
        target_page_id: { type: string, format: uuid }
        target_block_id: { type: string, format: uuid }
    BlockRedirect:
      type: object
      required:
        - source_page_id
        - source_block_id
        - target_page_id
        - target_block_id
        - target_page_title
        - target_current_slug
        - created_by
        - created_at
      properties:
        source_page_id: { type: string, format: uuid }
        source_block_id: { type: string, format: uuid }
        target_page_id: { type: string, format: uuid }
        target_block_id: { type: string, format: uuid }
        target_page_title: { type: string }
        target_current_slug: { type: string }
        created_by: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
    BlockRedirectList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/BlockRedirect" }
    PublishRevisionRequest:
      type: object
      required: [ast]
      properties:
        expected_revision_id:
          type: string
          format: uuid
          description: |
            发布基线（乐观锁）：必须等于页面当前 Revision。
            首发布不传（要求页面尚无 Revision）；不一致返回 409 stale_revision。
        working_document_id:
          type: string
          format: uuid
          description: |
            可选的 active WorkingDocument。提供时，发布 Revision 与工作副本换基
            在同一事务完成；文档基线过期时返回 409 stale_revision。
        ast:
          type: object
          additionalProperties: true
          description: |
            Typed Block AST v1 文档（权威 Schema：contracts/schemas/ast/v1/ast.schema.json）。
            必须是 JSON 对象，type='document' 且 schema_version=1。
        summary:
          type: string
          description: 修改摘要
          example: 初版
        is_minor:
          type: boolean
          description: 小修改标志
          default: false
    PageWithContent:
      type: object
      required: [page, content, via_alias]
      description: 阅读端点响应：落地页元信息 + 当前 Revision 内容（渲染直连 ContentSnapshot）
      properties:
        page:
          $ref: "#/components/schemas/Page"
        content:
          nullable: true
          description: 当前 Revision 内容；页面已创建但未发布过为 null
          $ref: "#/components/schemas/PageContent"
        via_alias:
          type: boolean
          description: 是否经由 page_alias（旧标题别名）解析命中
        alias_title:
          type: string
          description: via_alias=true 时回显客户端请求的标题原文
        redirect:
          type: object
          nullable: true
          description: >-
            请求命中重定向源时出现。resolved=true 时 page/content 是本地落地页；
            unresolved 或 interwiki 终点返回 resolved=false、content=null，并由 target
            给出可治理的目标。
          required: [from_page_id, from_title, target, resolved, hops]
          properties:
            from_page_id:
              type: string
              format: uuid
              description: 重定向源页面 ID
            from_title:
              type: string
              description: 重定向源页面显示标题
            target:
              $ref: "#/components/schemas/PageRedirectTarget"
            resolved:
              type: boolean
            hops:
              type: integer
              minimum: 1
...
    RevisionListPage:
      allOf:
        - $ref: "#/components/schemas/CursorPage"
        - type: object
          required: [items, next_cursor]
          description: Revision 历史列表（created_at DESC, id DESC 游标分页）
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Revision"
            next_cursor:
              type: string
              nullable: true
              description: 下一页游标；null 表示没有更多
...
  responses:
    BadRequest:
      description: 请求格式错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: 未认证
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: 已认证但无权限
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Gone:
      description: 资源曾存在但已删除（如软删除页面、重定向目标已删除）
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnprocessableEntity:
      description: 请求语义可理解但无法处理（如重定向环/重定向链过深）
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: 并发冲突（含陈旧基线、幂等键冲突）
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Internal:
      description: 服务端内部错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"